pulse_bengaluru_agent-parent/requirements.txt
google-cloud-aiplatform
google-adk
flask